Separate repository interface into monaxia-repository
