+++ backend/monaxia-job/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "monaxia-job"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
