+++ backend/monaxia/Cargo.toml
@@ -29,3 +29,4 @@ tower-http = { version = "0.4.3", features = ["trace"] }
 
 monaxia-data = { workspace = true }
 monaxia-db = { workspace = true }
+monaxia-repository = { workspace = true }
