+++ backend/monaxia/src/cli/migrate.rs
@@ -1,4 +1,4 @@
-use crate::{repository::Container, web::state::AppState};
+use crate::web::state::AppState;
 
 use std::{
     env::{current_dir, var as env_var, VarError},
@@ -8,6 +8,7 @@ use std::{
 
 use anyhow::{bail, Context, Result};
 use clap::Parser;
+use monaxia_repository::Container;
 use once_cell::sync::Lazy;
 use time::{
     format_description::FormatItem, macros::format_description, OffsetDateTime, PrimitiveDateTime,
